# src/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dockore",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/networks": {
      "get": {
        "tags": [
          "networks"
        ],
        "summary": "List Networks",
        "operationId": "list_networks_api_networks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NetworkItem"
                  },
                  "type": "array",
                  "title": "Response List Networks Api Networks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "networks"
        ],
        "summary": "Create Network",
        "operationId": "create_network_api_networks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetworkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "networks"
        ],
        "summary": "Delete Networks",
        "operationId": "delete_networks_api_networks_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/networks/{id}": {
      "get": {
        "tags": [
          "networks"
        ],
        "summary": "Get Network",
        "operationId": "get_network_api_networks__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/networks/{id}/connect": {
      "post": {
        "tags": [
          "networks"
        ],
        "summary": "Connect Network",
        "operationId": "connect_network_api_networks__id__connect_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/networks/{id}/disconnect": {
      "post": {
        "tags": [
          "networks"
        ],
        "summary": "Disconnect Network",
        "operationId": "disconnect_network_api_networks__id__disconnect_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisconnectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsData"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConnectRequest": {
        "properties": {
          "container_id": {
            "type": "string",
            "title": "Container Id"
          },
          "ipv4_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipv4 Address"
          }
        },
        "type": "object",
        "required": [
          "container_id"
        ],
        "title": "ConnectRequest"
      },
...
      "ContainerItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "image": {
            "$ref": "#/components/schemas/ImageItem"
          },
          "create_time": {
            "type": "string",
            "title": "Create Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "tty": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tty"
          },
          "interactive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interactive"
          },
          "network": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContainerNetwork"
              },
              {
                "type": "null"
              }
            ]
          },
          "mounts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Mount"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mounts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "image",
          "create_time",
          "status"
        ],
        "title": "ContainerItem"
      },
      "ContainerNetwork": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip",
            "default": ""
          },
          "prefix": {
            "type": "integer",
            "title": "Prefix",
            "default": 0
          },
          "gateway": {
            "type": "string",
            "title": "Gateway",
            "default": ""
          },
          "mac_address": {
            "type": "string",
            "title": "Mac Address",
            "default": ""
          },
          "ports": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PortMapping"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ports"
          }
        },
        "type": "object",
        "title": "ContainerNetwork"
      },
...
      "DeleteResult": {
        "properties": {
          "failed": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Failed",
            "default": {}
          }
        },
        "type": "object",
        "title": "DeleteResult",
        "description": "Failed ids mapped to their error message; empty means fully succeeded."
      },
      "DisconnectRequest": {
        "properties": {
          "container_id": {
            "type": "string",
            "title": "Container Id"
          }
        },
        "type": "object",
        "required": [
          "container_id"
        ],
        "title": "DisconnectRequest"
      },
...
      "ExposedPort": {
        "properties": {
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol"
          }
        },
        "type": "object",
        "required": [
          "port",
          "protocol"
        ],
        "title": "ExposedPort"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdsRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "IdsRequest"
      },
      "ImageItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": ""
          },
          "create_time": {
            "type": "string",
            "title": "Create Time"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "default": 0
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "tty": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tty"
          },
          "interactive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interactive"
          },
          "architecture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Architecture"
          },
          "os": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          },
          "ports": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ExposedPort"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ports"
          }
        },
        "type": "object",
        "required": [
          "id",
          "create_time"
        ],
        "title": "ImageItem"
      },
...
      "Mount": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "driver": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": ""
          },
          "src": {
            "type": "string",
            "title": "Src",
            "default": ""
          },
          "dest": {
            "type": "string",
            "title": "Dest",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Mount"
      },
      "NetworkCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "driver": {
            "type": "string",
            "title": "Driver"
          },
          "attachable": {
            "type": "boolean",
            "title": "Attachable",
            "default": true
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Options",
            "default": []
          },
          "subnet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subnet"
          },
          "gateway": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway"
          },
          "ip_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Range"
          }
        },
        "type": "object",
        "required": [
          "name",
          "driver"
        ],
        "title": "NetworkCreate"
      },
      "NetworkItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "driver": {
            "type": "string",
            "title": "Driver",
            "default": ""
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "create_time": {
            "type": "string",
            "title": "Create Time"
          },
          "container_num": {
            "type": "integer",
            "title": "Container Num",
            "default": 0
          },
          "subnet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subnet"
          },
          "gateway": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway"
          },
          "ip_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Range"
          },
          "ipam_driver": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipam Driver"
          },
          "internal": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal"
          },
          "attachable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachable"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "containers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ContainerItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Containers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "create_time"
        ],
        "title": "NetworkItem"
      },
      "Option": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "Option"
      },
      "PortMapping": {
        "properties": {
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "tcp"
          },
          "listen_ip": {
            "type": "string",
            "title": "Listen Ip",
            "default": "0.0.0.0"
          },
          "listen_port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Listen Port"
          }
        },
        "type": "object",
        "required": [
          "port",
          "listen_port"
        ],
        "title": "PortMapping"
      },
...
      "SettingsData": {
        "properties": {
          "docker_host": {
            "type": "string",
            "title": "Docker Host",
            "default": "unix:///var/run/docker.sock"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "SettingsData"
      },
      "SettingsUpdate": {
        "properties": {
          "docker_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docker Host"
          }
        },
        "type": "object",
        "title": "SettingsUpdate"
      },
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "StatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}